moved logic outside of action file. Moved scss to be alphabetical
