+++ src/Components/BidTracker/BidTracker.jsx
@@ -6,7 +6,9 @@ import ExportButton from 'Components/ExportButton';
 import SearchAsClientButton from 'Components/BidderPortfolio/SearchAsClientButton/SearchAsClientButton';
 import SelectForm from 'Components/SelectForm';
 import { BID_STATUS_ORDER } from 'Constants/BidStatuses';
-import { downloadBidlistData, getBidListStats } from 'actions/bidList';
+import { DRAFT_PROP, SUBMITTED_PROP } from 'Constants/BidData';
+import { downloadBidlistData } from 'actions/bidList';
+import { getBidListStats } from 'utilities';
 import { BID_LIST, NOTIFICATION_LIST, USER_PROFILE } from '../../Constants/PropTypes';
 import BidTrackerCardList from './BidTrackerCardList';
 import ProfileSectionTitle from '../ProfileSectionTitle';
@@ -95,8 +97,8 @@ class BidTracker extends Component {
     const cdoEmail = get(userProfile, 'cdo.email');
 
     const sortedBids = this.getSortedBids();
-    const draftBids = getBidListStats(bidList.results, 'draft', true);
-    const submittedBids = getBidListStats(bidList.results, 'submitted', true);
+    const draftBids = getBidListStats(bidList.results, DRAFT_PROP, true);
+    const submittedBids = getBidListStats(bidList.results, SUBMITTED_PROP, true);
     return (
       <div className="usa-grid-full profile-content-inner-container bid-tracker-page">
         <BackButton />
