+++ src/sass/_bidTracker.scss
@@ -96,8 +96,8 @@ $register-icon-offset: 44%;
       margin-right: 20px;
     }
     @media screen and (min-width: 1200px) and (max-width: 1299px) {
-      margin: 5px 0 0 0;
       font-size: 12px;
+      margin: 5px 0 0 0;
       .bid-stat {
         font-weight: normal;
       }
@@ -106,8 +106,8 @@ $register-icon-offset: 44%;
       }
     }
     @media screen and (min-width: 1300px) and (max-width: 1400px) {
-      margin: 3px 10px 0 0;
       font-size: 14px;
+      margin: 3px 10px 0 0;
       div:first-child {
         margin-right: 10px;
       }
